Change explicit targeting config types
This switches the explicit targeting config to be driven by self-contained rules
which themselves define how to extract one or more feature names out of the request.
The rules only need to implement a trait in order to be evaluated by the system.

This means that
1. Explicit targeting operates on the request as a whole as input
2. Adapters can supply specific configuration applicable to the host
3. Adapters can implement custom explicit targeting rules which the core
   doesn't support in case they need to

The implict targeting rules will likely be a similar system, except instead of
extracting feature names, they will return a bool indicating whether the rule
matched on the request or not.